Refactor a bit the poisoners
