+++ poisoners/NBTNS.py
@@ -14,8 +14,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-import socket
-import settings
+
 import fingerprint
 
 from packets import NBT_Ans
@@ -26,20 +25,15 @@ from utils import *
 def Validate_NBT_NS(data):
 	if settings.Config.AnalyzeMode:
 		return False
-
-	if NBT_NS_Role(data[43:46]) == "File Server":
+	elif NBT_NS_Role(data[43:46]) == "File Server":
 		return True
-
-	if settings.Config.NBTNSDomain:
+	elif settings.Config.NBTNSDomain:
 		if NBT_NS_Role(data[43:46]) == "Domain Controller":
 			return True
-
-	if settings.Config.Wredirect:
+	elif settings.Config.Wredirect:
 		if NBT_NS_Role(data[43:46]) == "Workstation/Redirector":
 			return True
-
-	else:
-		return False
+	return False
 
 # NBT_NS Server class.
 class NBTNS(BaseRequestHandler):
@@ -54,19 +48,14 @@ class NBTNS(BaseRequestHandler):
 			return None
 
 		if data[2:4] == "\x01\x10":
-
+			Finger = None
 			if settings.Config.Finger_On_Off:
 				Finger = fingerprint.RunSmbFinger((self.client_address[0],445))
-			else:
-				Finger = None
 
-			# Analyze Mode
-			if settings.Config.AnalyzeMode:
+			if settings.Config.AnalyzeMode:  # Analyze Mode
 				LineHeader = "[Analyze mode: NBT-NS]"
 				print color("%s Request by %s for %s, ignoring" % (LineHeader, self.client_address[0], Name), 2, 1)
-
-			# Poisoning Mode
-			else:
+			else:  # Poisoning Mode
 				Buffer = NBT_Ans()
 				Buffer.calculate(data)
 				socket.sendto(str(Buffer), self.client_address)
